requirements.txt
numpy
matplotlib
pandas
plotly
python-dotenv